Fix int64→int32 pointer stride mismatch in packed storage
pack_tensor stored packed weights as int64 (8 bytes/element) but the CUDA
kernel reads them as uint32 (4 bytes/element). Each int64 was being split
into two uint32 reads, corrupting every other packed word.

Changed pack_tensor to return int32, and the CUDA launch wrapper to read
int32_t* — stride_words now matches between Python and CUDA.

+++ kernels/packed_ternary/__init__.py
@@ -244,7 +244,9 @@ def pack_tensor(fp32_tensor: torch.Tensor, gamma: float = 1.0) -> torch.Tensor:
             chunk += [0] * (16 - len(chunk))
             packed[r, w] = pack16_reference(chunk)
 
-    return torch.from_numpy(packed.astype(np.int64))
+    # Must be int32 so the CUDA kernel can read stride_words uint32 entries
+    # per row without pointer-arithmetic mismatch (int64 → uint32 halves data).
+    return torch.from_numpy(packed.astype(np.int32))
 
 
 def unpack_tensor(packed: torch.Tensor, rows: int, cols: int, gamma: float = 1.0) -> torch.Tensor:
